Usage of atomic transaction block in middleware file.
Should help with #26 .

Difficult to test for sure. The change now wraps the creation of the
silk response model in a atomic transaction block. If this still fails
it should gracefully be caught with the IntegrityError and continue on.

It should be noted that middleware classes in Django fall outside the
transactions.