Fixed a bug that the profile tab could not be opened when the source code contained Japanese.
There was a bug that the profile tab could not be opened when the source code contained Japanese. The error message is as follows.

'cp932' codec can't decode byte 0x88 in position 19: illegal multibyte sequence

Obviously, it is due to a character code problem.
Basically, the source code should be written in UTF-8, so I explicitly specified to read it in utf-8.